Document paginated instrument and stock price search APIs

+++ src/BorsdataMcp/Tools/InstrumentLookup.cs
@@ -6,15 +6,15 @@ namespace BorsdataMcp.Tools;
 // GetShortHoldings) or parse a user-supplied comma-separated instrumentIds filter.
 internal static class InstrumentLookup
 {
-    public static Dictionary<int, (string Name, string Ticker)> BuildIndex(JsonNode? instrumentsRoot)
+    public static Dictionary<long, (string Name, string Ticker)> BuildIndex(JsonNode? instrumentsRoot)
     {
         // Börsdata wraps the array in an envelope object, e.g. { "instruments": [ ... ] }.
         var all = (instrumentsRoot as JsonObject)?["instruments"] as JsonArray ?? [];
-        var index = new Dictionary<int, (string, string)>();
+        var index = new Dictionary<long, (string, string)>();
 
         foreach (var node in all.OfType<JsonObject>())
         {
-            if (node["insId"] is JsonValue idValue && idValue.TryGetValue(out int insId) &&
+            if (node["insId"] is JsonValue idValue && idValue.TryGetValue(out long insId) &&
                 node["name"] is JsonValue nameValue && nameValue.TryGetValue(out string? name) &&
                 node["ticker"] is JsonValue tickerValue && tickerValue.TryGetValue(out string? ticker))
             {
@@ -25,14 +25,14 @@ internal static class InstrumentLookup
         return index;
     }
 
-    public static HashSet<int>? ParseIds(string? commaSeparatedIds)
+    public static HashSet<long>? ParseIds(string? commaSeparatedIds)
     {
         if (string.IsNullOrWhiteSpace(commaSeparatedIds))
             return null;
 
         return commaSeparatedIds
             .Split(',')
-            .Select(s => int.TryParse(s.Trim(), out var id) ? id : (int?)null)
+            .Select(s => long.TryParse(s.Trim(), out var id) ? id : (long?)null)
             .Where(id => id is not null)
             .Select(id => id!.Value)
             .ToHashSet();
